requirements.txt
opencv_python==4.2.0.34
pandas==1.0.3
seaborn==0.11.0
torch==1.5.0
tabulate==0.8.7
scipy==1.4.1
~umpy==1.18.1
moviepy==1.0.3
hdf5storage==0.1.15
h5py==2.10.0
matplotlib==3.0.3
tqdm==4.45.0
torchvision==0.6.0
ipdb==0.13.8
numpy==1.20.3
Pillow==8.2.0
PyYAML==5.4.1
scikit_learn==0.24.2
tensorboardX==2.2
